Share app shell and source plugin wiring
Extract common cloud/local app shell pieces and centralize each app's source plugin list so routes and shells do not repeat the same imports and arrays.

Preserves app-specific loading states, MCP stdio settings, cloud org UI, and route behavior.
