+++ apps/cloud/src/routes/sources.$namespace.tsx
@@ -1,10 +1,7 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { SourceDetailPage } from "@executor/react/pages/source-detail";
-import { openApiSourcePlugin } from "@executor/plugin-openapi/react";
-import { mcpSourcePlugin } from "@executor/plugin-mcp/react";
-import { graphqlSourcePlugin } from "@executor/plugin-graphql/react";
+import { sourcePlugins } from "../web/source-plugins";
 
-const sourcePlugins = [openApiSourcePlugin, mcpSourcePlugin, graphqlSourcePlugin];
 
 export const Route = createFileRoute("/sources/$namespace")({
   component: () => {
