+++ apps/cloud/src/routes/sources.add.$pluginKey.tsx
@@ -1,11 +1,8 @@
 import { Schema } from "effect";
 import { createFileRoute } from "@tanstack/react-router";
 import { SourcesAddPage } from "@executor/react/pages/sources-add";
-import { openApiSourcePlugin } from "@executor/plugin-openapi/react";
-import { mcpSourcePlugin } from "@executor/plugin-mcp/react";
-import { graphqlSourcePlugin } from "@executor/plugin-graphql/react";
+import { sourcePlugins } from "../web/source-plugins";
 
-const sourcePlugins = [openApiSourcePlugin, mcpSourcePlugin, graphqlSourcePlugin];
 
 const SearchParams = Schema.standardSchemaV1(
   Schema.Struct({
