+++ apps/local/src/routes/sources.$namespace.tsx
@@ -1,18 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { SourceDetailPage } from "@executor/react/pages/source-detail";
-import { openApiSourcePlugin } from "@executor/plugin-openapi/react";
-import { createMcpSourcePlugin } from "@executor/plugin-mcp/react";
-
-const mcpSourcePlugin = createMcpSourcePlugin({ allowStdio: true });
-import { googleDiscoverySourcePlugin } from "@executor/plugin-google-discovery/react";
-import { graphqlSourcePlugin } from "@executor/plugin-graphql/react";
-
-const sourcePlugins = [
-  openApiSourcePlugin,
-  mcpSourcePlugin,
-  googleDiscoverySourcePlugin,
-  graphqlSourcePlugin,
-];
+import { sourcePlugins } from "../web/source-plugins";
 
 export const Route = createFileRoute("/sources/$namespace")({
   component: () => {
