+++ apps/local/src/routes/sources.add.$pluginKey.tsx
@@ -1,19 +1,7 @@
 import { Schema } from "effect";
 import { createFileRoute } from "@tanstack/react-router";
 import { SourcesAddPage } from "@executor/react/pages/sources-add";
-import { openApiSourcePlugin } from "@executor/plugin-openapi/react";
-import { createMcpSourcePlugin } from "@executor/plugin-mcp/react";
-
-const mcpSourcePlugin = createMcpSourcePlugin({ allowStdio: true });
-import { googleDiscoverySourcePlugin } from "@executor/plugin-google-discovery/react";
-import { graphqlSourcePlugin } from "@executor/plugin-graphql/react";
-
-const sourcePlugins = [
-  openApiSourcePlugin,
-  mcpSourcePlugin,
-  googleDiscoverySourcePlugin,
-  graphqlSourcePlugin,
-];
+import { sourcePlugins } from "../web/source-plugins";
 
 const SearchParams = Schema.standardSchemaV1(
   Schema.Struct({
